refactor(optimize): introduce 6-method branch optimization scaffolding
Split `optimize_unified` into `method_newton` and `method_brent` modules, each owning its solver functions and inner-loop constants. Shared infrastructure (metrics, contributions, evaluation, grid search, zero-branch logic, `run_optimize_mixed`) stays in `optimize_unified`.

Expand `BranchOptMethod` to six variants covering the two orthogonal axes of algorithm and reparameterization: `Brent`, `BrentSqrt`, `BrentLog`, `Newton`, `NewtonSqrt`, `NewtonLog`. Move the enum into `args.rs` beside `InitialGuessMode`, delete `optimize_method.rs`, and update the 13 import sites across the crate.

Set `BrentSqrt` as the target default to match v0. Variants without an implementation are wired as `todo!()` in dispatch and filled in by later commits on this branch.
