+++ packages/treetime/src/commands/optimize/__tests__/test_convergence/test_convergence_iterations.rs
@@ -1,7 +1,7 @@
 #[cfg(test)]
 mod tests {
   use crate::commands::ancestral::marginal::update_marginal;
-  use crate::commands::optimize::optimize_method::BranchOptMethod;
+  use crate::commands::optimize::args::BranchOptMethod;
   use crate::commands::optimize::optimize_unified::run_optimize_mixed;
   use crate::representation::payload::ancestral::GraphAncestral;
   use eyre::Report;
