+++ packages/treetime/src/commands/optimize/__tests__/test_gm_optimize.rs
@@ -110,7 +110,7 @@ mod tests {
     use crate::alphabet::alphabet::Alphabet;
     use crate::commands::ancestral::fitch::{compress_sequences, get_common_length};
     use crate::commands::ancestral::marginal::{initialize_marginal, update_marginal};
-    use crate::commands::optimize::optimize_method::BranchOptMethod;
+    use crate::commands::optimize::args::BranchOptMethod;
     use crate::commands::optimize::optimize_unified::{initial_guess_mixed, run_optimize_mixed};
     use crate::commands::optimize::run::{apply_damping, collect_optimize_partitions, save_branch_lengths};
     use crate::gtr::get_gtr::{JC69Params, jc69};
