+++ packages/treetime/src/commands/optimize/mod.rs
@@ -2,11 +2,12 @@
 mod __tests__;
 
 pub mod args;
+pub mod method_brent;
+pub mod method_newton;
 pub mod optimize_dense;
 mod optimize_dense_eval;
 mod optimize_eval;
 pub mod optimize_indel;
-pub mod optimize_method;
 pub mod optimize_sparse;
 mod optimize_sparse_eval;
 pub mod optimize_unified;
